// src/app/shared/models/registerUser.ts
export class RegisterUser{
  id: number
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  email: string;
  displayName: string;
  description: string;
  socialMediaLink: string;
  phoneNumber: string;
  address: string;
  city: string;
  country: string;
  creditCardNumber: string;

  username: string;
  password1: string;
  password2: string;

  artistId: number;
}